Refactored common RNG class into include random_quda.h Left curand and hiprand implementations of uniform:: and normal:: in include/targets/<arch>/random_quda_impl.h Adapted gauge and spinor algorithms to use uniform<T>::rand() instead of Random<T> since they were implemented identically Removed Random<T> and left uniform<T>::rand() in place to reduce duplication/boilerplate. Summit tests pass with the exclption of blas_test_wilson_fill (which also failed before refactor) (@maddyscientist is aware)
