+++ include/kernels/gauge_random.cuh
@@ -45,8 +45,8 @@ namespace quda {
     real rand1[4], rand2[4], phi[4], radius[4], temp1[4], temp2[4];
 
     for (int i = 0; i < 4; ++i) {
-      rand1[i] = Random<real>(localState);
-      rand2[i] = Random<real>(localState);
+      rand1[i] = uniform<real>::rand(localState);
+      rand2[i] = uniform<real>::rand(localState);
     }
 
     for (int i = 0; i < 4; ++i) {
